frontend-ui/requirements.txt
streamlit>=1.39.0
httpx==0.28.1
python-dotenv==1.1.0
pandas>=2.0.0